+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Add CI workflow to check for upstream Alpine updated.
+- Add supervisord to combine all relevant logs and publish to docker logs.
 ### Changed
 - Improve documentation.
 - Change PHP CS ruleset such that global objects are always imported.